+++ server/src/external/stripe/customers/utils/buildIdempotencyKey.ts
@@ -1,6 +1,9 @@
 import { hashString } from "@autumn/shared";
+import { AUTUMN_STRIPE_IDEMPOTENCY_PREFIX } from "@/external/stripe/common/autumnStripeIdempotency";
 import type { AutumnContext } from "@/honoUtils/HonoEnv";
 
+/** Prefixed so webhooks can attribute the event to Autumn; the hash still
+ * dedupes concurrent creates within the same 5s bucket. */
 export const buildStripeCustomerIdempotencyKey = ({
 	ctx,
 	customerId,
@@ -9,7 +12,7 @@ export const buildStripeCustomerIdempotencyKey = ({
 	customerId: string;
 }): string => {
 	const { org, env } = ctx;
-	return hashString(
+	return `${AUTUMN_STRIPE_IDEMPOTENCY_PREFIX}customer:${hashString(
 		`stripe-create-cus:${customerId}:${org.id}:${env}:${Math.floor(Date.now() / 5000)}`,
-	);
+	)}`;
 };
